samples/math-calculator-agent/pyproject.toml
[project]
name = "math-calculator-agent"
version = "0.0.1"
description = "Add your description here"
authors = [{ name = "John Doe", email = "john.doe@myemail.com" }]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
dependencies = [
    "llama-index-llms-openai>=0.2.2",
    "uipath-llamaindex>=0.0.11",
]
